Fix タイムアウト判定。

+++ src/cosmic/universe.rs
@@ -10,8 +10,8 @@ pub struct Universe {
     pub game: Game,
     /// 対話モード
     pub dialogue_mode: bool,
-    /// 読みの最大深さ。最大で 255 だぜ☆（＾～＾） そんな要らんだろ☆（＾～＾）
-    pub option_max_depth: u8,
+    /// 読みの最大深さ。
+    pub option_max_depth: usize,
     /// 思考時間の最小秒☆（＾～＾）
     pub option_min_think_sec: u64,
     /// 思考時間の最大秒☆（＾～＾）
@@ -19,7 +19,7 @@ pub struct Universe {
     /// 盤面をカバーする利きの多さ☆（＾～＾）1000分率☆（＾～＾）
     pub option_board_coverage_weight: i32,
     /// 諦めない深さ☆（＾～＾）読み終わるまで、思考時間を無視するぜ☆（＾～＾）max_depth - 1 より小さくしろだぜ☆（＾～＾）
-    pub option_depth_not_to_give_up: u8,
+    pub option_depth_not_to_give_up: usize,
     /// 駒割の重み☆（＾～＾）1000分率☆（＾～＾）
     pub option_komawari_weight: i32,
     /// 成りの重み☆（＾～＾）1000分率☆（＾～＾）
